# frontend/src/generated/backend-openapi.json
{
  "openapi": "3.1.0",
  "info": {
    "title": "Starlib API",
    "description": "Backend API for Starlib music management",
    "version": "0.1.0"
  },
  "paths": {
...
    "/api/rekordbox/usb/eject": {
      "post": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Eject Usb Device",
        "description": "Safely eject a mounted Rekordbox USB/SD export.",
        "operationId": "eject_usb_device_api_rekordbox_usb_eject_post",
        "parameters": [
          {
            "name": "device",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "USB device id",
              "title": "Device"
            },
            "description": "USB device id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/EjectResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/status": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Status",
        "description": "Report whether the selected Rekordbox source is reachable.",
        "operationId": "get_status_api_rekordbox_status_get",
        "parameters": [
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/RekordboxStatus"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/playlists": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Playlists",
        "operationId": "get_playlists_api_rekordbox_playlists_get",
        "parameters": [
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/PlaylistsResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/playlists/{playlist_id}/tracks": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Playlist Tracks",
        "operationId": "get_playlist_tracks_api_rekordbox_playlists__playlist_id__tracks_get",
        "parameters": [
          {
            "name": "playlist_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Playlist Id"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/tracks/{track_id}/artwork": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Track Artwork",
        "description": "Serve the cached artwork JPEG for a track.",
        "operationId": "get_track_artwork_api_rekordbox_tracks__track_id__artwork_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "small",
            "in": "query",
            "required": false,
            "schema": {
              "type": "boolean",
              "default": true,
              "title": "Small"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/tracks/{track_id}/waveform": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Track Waveform",
        "description": "Serve a track's raw ANLZ waveform bytes for the frontend canvas.\n\nPreview variants span the whole track: ``color`` = PWV4 (1200 x 6 bytes),\n``blue`` = PWAV (400 x 1). Detail variants carry ~150 columns/second for\nzoomed playback: ``color_detail`` = PWV5 (2 bytes/column, ``.EXT``),\n``blue_detail`` = PWV3 (1 byte/column, ``.DAT``).",
        "operationId": "get_track_waveform_api_rekordbox_tracks__track_id__waveform_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          },
          {
            "name": "variant",
            "in": "query",
            "required": false,
            "schema": {
              "type": "string",
              "pattern": "^(color|blue|color_detail|blue_detail)$",
              "default": "color",
              "title": "Variant"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/tracks/{track_id}/analysis": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Track Analysis",
        "description": "Return a track's beatgrid, phrase sections and cues as JSON.\n\nDrives the zoomed player overlay: beat ticks, the phrase band and cue\nmarkers. ``sections`` is ``null`` when the track has no phrase analysis;\nbeatgrid/cues degrade to empty lists rather than 404 when analysis is absent.",
        "operationId": "get_track_analysis_api_rekordbox_tracks__track_id__analysis_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TrackAnalysisResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/tracks/{track_id}/audio": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Stream Track Audio",
        "description": "Stream a track's audio file from a mounted USB export.\n\nLocal-install tracks play through the metadata audio endpoint by their\nabsolute path; USB tracks live on the device, so their device-relative path\nis resolved here (within the mount) and streamed with the same\ntranscode/range handling.",
        "operationId": "stream_track_audio_api_rekordbox_tracks__track_id__audio_get",
        "parameters": [
          {
            "name": "track_id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "title": "Track Id"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": true,
            "schema": {
              "type": "string",
              "description": "USB device id",
              "title": "Device"
            },
            "description": "USB device id"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {}
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    },
    "/api/rekordbox/tracks": {
      "get": {
        "tags": [
          "rekordbox"
        ],
        "summary": "Get Tracks",
        "operationId": "get_tracks_api_rekordbox_tracks_get",
        "parameters": [
          {
            "name": "limit",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "integer",
                  "maximum": 10000,
                  "minimum": 1
                },
                {
                  "type": "null"
                }
              ],
              "title": "Limit"
            }
          },
          {
            "name": "device",
            "in": "query",
            "required": false,
            "schema": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "null"
                }
              ],
              "description": "USB device id; omit for the local install",
              "title": "Device"
            },
            "description": "USB device id; omit for the local install"
          }
        ],
        "responses": {
          "200": {
            "description": "Successful Response",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/TracksResponse"
                }
              }
            }
          },
          "422": {
            "description": "Validation Error",
            "content": {
              "application/json": {
                "schema": {
                  "$ref": "#/components/schemas/HTTPValidationError"
                }
              }
            }
          }
        }
      }
    }
  },
  "components": {
    "schemas": {
...
      "BeatModel": {
        "properties": {
          "beat": {
            "type": "integer",
            "title": "Beat"
          },
          "bpm": {
            "type": "number",
            "title": "Bpm"
          },
          "timeMs": {
            "type": "integer",
            "title": "Timems"
          }
        },
        "type": "object",
        "required": [
          "beat",
          "bpm",
          "timeMs"
        ],
        "title": "BeatModel"
      },
...
      "CueModel": {
        "properties": {
          "type": {
            "type": "string",
            "title": "Type"
          },
          "index": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Index"
          },
          "timeMs": {
            "type": "integer",
            "title": "Timems"
          },
          "color": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Color"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "outMs": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Outms"
          }
        },
        "type": "object",
        "required": [
          "type",
          "timeMs"
        ],
        "title": "CueModel"
      },
...
      "EjectResponse": {
        "properties": {
          "ok": {
            "type": "boolean",
            "title": "Ok"
          }
        },
        "type": "object",
        "required": [
          "ok"
        ],
        "title": "EjectResponse"
      },
...
      "HTTPValidationError": {
        "properties": {
          "detail": {
            "items": {
              "$ref": "#/components/schemas/ValidationError"
            },
            "type": "array",
            "title": "Detail"
          }
        },
        "type": "object",
        "title": "HTTPValidationError"
      },
...
      "PlaylistsResponse": {
        "properties": {
          "playlists": {
            "items": {
              "$ref": "#/components/schemas/RekordboxPlaylist"
            },
            "type": "array",
            "title": "Playlists"
          }
        },
        "type": "object",
        "required": [
          "playlists"
        ],
        "title": "PlaylistsResponse"
      },
...
      "RekordboxPlaylist": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "name": {
            "type": "string",
            "title": "Name"
          },
          "parent_id": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Parent Id"
          },
          "is_folder": {
            "type": "boolean",
            "title": "Is Folder"
          },
          "is_smart": {
            "type": "boolean",
            "title": "Is Smart"
          },
          "track_count": {
            "type": "integer",
            "title": "Track Count"
          }
        },
        "type": "object",
        "required": [
          "id",
          "name",
          "is_folder",
          "is_smart",
          "track_count"
        ],
        "title": "RekordboxPlaylist"
      },
      "RekordboxStatus": {
        "properties": {
          "available": {
            "type": "boolean",
            "title": "Available"
          },
          "reason": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Reason"
          }
        },
        "type": "object",
        "required": [
          "available"
        ],
        "title": "RekordboxStatus"
      },
      "RekordboxTrack": {
        "properties": {
          "id": {
            "type": "string",
            "title": "Id"
          },
          "title": {
            "type": "string",
            "title": "Title"
          },
          "artist": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Artist"
          },
          "album": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Album"
          },
          "genre": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Genre"
          },
          "bpm": {
            "anyOf": [
              {
                "type": "number"
              },
              {
                "type": "null"
              }
            ],
            "title": "Bpm"
          },
          "key": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Key"
          },
          "duration_seconds": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Duration Seconds"
          },
          "file_path": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "File Path"
          },
          "comment": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Comment"
          },
          "soundcloud_id": {
            "anyOf": [
              {
                "type": "integer"
              },
              {
                "type": "null"
              }
            ],
            "title": "Soundcloud Id"
          },
          "date_added": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Date Added"
          },
          "release_date": {
            "anyOf": [
              {
                "type": "string"
              },
              {
                "type": "null"
              }
            ],
            "title": "Release Date"
          },
          "has_artwork": {
            "type": "boolean",
            "title": "Has Artwork",
            "default": false
          },
          "has_waveform": {
            "type": "boolean",
            "title": "Has Waveform",
            "default": false
          }
        },
        "type": "object",
        "required": [
          "id",
          "title"
        ],
        "title": "RekordboxTrack"
      },
...
      "SectionModel": {
        "properties": {
          "kind": {
            "type": "string",
            "title": "Kind"
          },
          "label": {
            "type": "string",
            "title": "Label"
          },
          "startMs": {
            "type": "integer",
            "title": "Startms"
          },
          "endMs": {
            "type": "integer",
            "title": "Endms"
          }
        },
        "type": "object",
        "required": [
          "kind",
          "label",
          "startMs",
          "endMs"
        ],
        "title": "SectionModel"
      },
...
      "TrackAnalysisResponse": {
        "properties": {
          "beatgrid": {
            "items": {
              "$ref": "#/components/schemas/BeatModel"
            },
            "type": "array",
            "title": "Beatgrid"
          },
          "sections": {
            "anyOf": [
              {
                "items": {
                  "$ref": "#/components/schemas/SectionModel"
                },
                "type": "array"
              },
              {
                "type": "null"
              }
            ],
            "title": "Sections"
          },
          "cues": {
            "items": {
              "$ref": "#/components/schemas/CueModel"
            },
            "type": "array",
            "title": "Cues"
          }
        },
        "type": "object",
        "required": [
          "beatgrid",
          "cues"
        ],
        "title": "TrackAnalysisResponse"
      },
...
      "TracksResponse": {
        "properties": {
          "tracks": {
            "items": {
              "$ref": "#/components/schemas/RekordboxTrack"
            },
            "type": "array",
            "title": "Tracks"
          }
        },
        "type": "object",
        "required": [
          "tracks"
        ],
        "title": "TracksResponse"
      },
...
      "ValidationError": {
        "properties": {
          "loc": {
            "items": {
              "anyOf": [
                {
                  "type": "string"
                },
                {
                  "type": "integer"
                }
              ]
            },
            "type": "array",
            "title": "Location"
          },
          "msg": {
            "type": "string",
            "title": "Message"
          },
          "type": {
            "type": "string",
            "title": "Error Type"
          },
          "input": {
            "title": "Input"
          },
          "ctx": {
            "type": "object",
            "title": "Context"
          }
        },
        "type": "object",
        "required": [
          "loc",
          "msg",
          "type"
        ],
        "title": "ValidationError"
      }
    }
  }
}